Allure, Report Portal reports
