+++ pom.xml
@@ -11,13 +11,14 @@
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
         <maven.compiler.target>17</maven.compiler.target>
+        <aspectj.version>1.8.10</aspectj.version>
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>org.testng</groupId>
             <artifactId>testng</artifactId>
-            <version>RELEASE</version>
+            <version>7.1.0</version>
             <scope>test</scope>
         </dependency>
     <dependency>
@@ -35,8 +36,72 @@
             <artifactId>selenium-support</artifactId>
             <version>4.1.4</version>
         </dependency>
-
-
+        <dependency>
+            <groupId>log4j</groupId>
+            <artifactId>log4j</artifactId>
+            <version>1.2.17</version>
+        </dependency>
+        <dependency>
+            <groupId>io.qameta.allure</groupId>
+            <artifactId>allure-testng</artifactId>
+            <version>2.17.3</version>
+        </dependency>
+        <dependency>
+            <groupId>com.epam.reportportal</groupId>
+            <artifactId>agent-java-testng</artifactId>
+            <version>5.0.11</version>
+        </dependency>
     </dependencies>
+
+    <build>
+            <plugins>
+    <plugin>
+        <groupId>io.qameta.allure</groupId>
+        <artifactId>allure-maven</artifactId>
+        <version>2.10.0</version>
+    </plugin>
+    <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-surefire-plugin</artifactId>
+        <version>2.21.0</version>
+        <configuration>
+            <properties>
+                <property>
+                    <name>usedefaultlisteners</name>
+                    <value>false</value>
+                </property>
+                <property>
+                    <name>listener</name>
+                    <value>com.epam.reportportal.testng.ReportPortalTestNGListener</value>
+                </property>
+            </properties>
+            <testFailureIgnore>false</testFailureIgnore>
+            <argLine>
+                -javaagent:"${settings.localRepository}/org/aspectj/aspectjweaver/${aspectj.version}/aspectjweaver-${aspectj.version}.jar"
+            </argLine>
+            <systemProperties>
+                <property>
+                    <name> allure.results.directory </name>
+                    <value> ${project.build.directory}/allure-results </value>>
+                </property>
+            </systemProperties>
+            <systemPropertyVariables>
+                <allure.results.directory> ${project.build.directory}/allure-results </allure.results.directory>
+                <rp.endpoint>https://demo.reportportal.io</rp.endpoint>
+                <rp.uuid>7c49fb66-de54-4069-b5b6-efca3836feb3</rp.uuid>
+                <rp.launch>Rubber ducks tests</rp.launch>
+                <rp.project>glebemelianov_personal</rp.project>
+            </systemPropertyVariables>
+        </configuration>
+        <dependencies>
+            <dependency>
+                <groupId>org.aspectj</groupId>
+                <artifactId>aspectjweaver</artifactId>
+                <version>${aspectj.version}</version>
+            </dependency>
+        </dependencies>
+    </plugin>
+            </plugins>
+    </build>
 
 </project>
