+++ src/test/java/resources/log4j.properties.properties
@@ -1,0 +1,4 @@
+log4j.rootLogger=INFO,console
+log4j.appender.console=org.apache.log4j.ConsoleAppender
+log4j.appender.console.layout=org.apache.log4j.PatternLayout
+log4j.appender.console.layout.ConversionPattern= %p %c: %m%n
